Moved components to components/ and logic to helpers/
