+++ src/components/ResultSegment.jsx
