+++ src/routes/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
-import nanoid from 'nanoid';
+import nanoid from 'nanoid/non-secure';
 import PropTypes from 'prop-types';
 
 import data from './data';
